Refactored server routes
